Adapt the llvm-prepo code to pstore API changes and add repo2obj support for internal fixups.
This change switches the StringTable code so that it now uses pstore::sstring_view<> exclusively rather than the policy-based templates that were previously necessary. The main per-fragment loop is now cut in two: the first phase records the OutputSection and size of each section prior to appending data to it. This is then used by the append function to generate the relocations that correspond to each internal fixup.