Modify the pstore string set so that it now contains "indirect" references to the strings.
The repo linker's shadow-memory will use the address in the ticket records as a pointer to the individual symbol records that it is building and using. To avoid using too much virtual memory, I need these pointers to cluster tightly together. To achieve this, the string set has been modified so that it stores pointers to strings rather than storing the strings directly. I then arrange for those pointers to be written in a group and then followed by the string bodies.