Eliminate the fragment_type/section_type distinction.
This change removes the distinction between fragment_type (which was the complete collection of valid section kinds) and section_type (which was the collection of sections that aren't related to the repository itself). This makes sense because it's important for repo consumers to understand the meaning of the various sections anyway. It also simplifies the code a little and allows for more symmetry, which is always a good thing. The fragment::dependents() function is removed in favor of fragment::atp<>() which can access any section.