Rename ticket_ -> compilation_member. Add an extent field field.
The ticket_member struct is renamed "compilation_member" (part of a
broader renaming so that we eliminate the excessive use of the term
"ticket" for almost everything!

To eliminate the need for the consumers to so a search of the fragment
index if they want to access the fragment associated with a
compilation_member, a new 'fext' (fragment extent) field is added to
that struct.